+++ src/docs/conf.py
@@ -8,9 +8,9 @@ author = u'Tom Oinn'
 # built documents.
 #
 # The short X.Y version.
-version = '2.4.3'
+version = '2.5.0'
 # The full version, including alpha/beta/rc tags.
-release = '2.4.3'
+release = '2.5.0'
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom